Fixes #272

+++ ECommerce.API/Interface/IProductUserRankRepository.cs
@@ -6,5 +6,5 @@ public interface IProductUserRankRepository : IAsyncRepository<ProductUserRank>
 {
     Task<ProductUserRank?> GetByProductUser(int productId, int userId, CancellationToken cancellationToken);
 
-    Task<int> GetBySumProduct(int productId, CancellationToken cancellationToken);
+    Task<double> GetBySumProduct(int productId, CancellationToken cancellationToken);
 }
